pyproject.toml
[project]
name = "castor_gallery"
description = "A static image gallery generator."
authors = [{name = "Daniel Williams", email ="mail@daniel-williams.co.uk"}]
repository = "https://github.com/transientlunatic/castor"
documentation = "https://transientlunatic.github.io/castor/"
readme = "README.md"

dynamic = [
	"version"
	]

click = "^8.1.8"

dependencies = [
"setuptools",
	     "mypy",
	     "pre-commit",
	     "tox",
	     "pillow (>=11.2.1,<12.0.0)",
]


[tool.setuptools_scm]
local_scheme = "no-local-version"

[tool.poetry]
version = "0.0.0"
packages = [
  {include = "castor_gallery", from = "python", to = "castor-gallery"},
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"

[project.scripts]
castor = "castor_gallery.__main__:castor"

[project.urls]
#"Source code" = "https://"

[tool.setuptools]
include-package-data = true
packages = [
	"python/castor_gallery",
]




[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["python/castor_gallery"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"



[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["python/castor_gallery"]


[tool.ruff.per-file-ignores]
"python/tests/*" = ["S101"]
